requirements.txt
asgiref==3.8.1
Django==5.0.6
django-mailbox==4.10.1
django-smart-selects==1.7.1
pillow==10.3.0
sqlparse==0.5.0
